+++ src/ff.rs
@@ -168,6 +168,8 @@ pub enum Error {
     FfNotSupported,
     /// Requested effect is not supported by device
     EffectNotSupported,
+    /// Can not play effect
+    FailedToPlay,
 }
 
 impl Error {
@@ -176,6 +178,7 @@ impl Error {
             Error::NotEnoughSpace => "not enough space for new effect",
             Error::FfNotSupported => "force feedback is not supported",
             Error::EffectNotSupported => "effect is not supported by device",
+            Error::FailedToPlay => "can't play effect"
         }
     }
 }