+++ CHANGELOG.md
@@ -16,6 +16,7 @@ v0.3.0 - unreleased
 - Rename `Button::Unknow` to `Button::Unknown`
 - `Gamepad::name(&self)` now returns `&str` instead of `&String`
 - Improved dead zone detection
+- `Effect::play(&self, u16)` now returns `Result<(), Error>`
 - Linux: Reduced memory usage
 
 ### Removed